Remove extension class and put methods on interface instead
