+++ src/NuGet.Clients/NuGet.PackageManagement.UI/Actions/PackagesConfigToPackageReferenceMigrator.cs
@@ -15,6 +15,7 @@ using NuGet.Packaging.Core;
 using NuGet.ProjectManagement;
 using NuGet.ProjectManagement.Projects;
 using NuGet.Protocol.Core.Types;
+using NuGet.VisualStudio.Internal.Contracts;
 
 namespace NuGet.PackageManagement.UI
 {
@@ -101,7 +102,7 @@ namespace NuGet.PackageManagement.UI
                     nuGetProject = await solutionManager.UpgradeProjectToPackageReferenceAsync(nuGetProject);
 
                     // Ensure we use the updated project for installing, and don't display preview or license acceptance windows.
-                    var projectContextInfo = await ProjectContextInfoExtensions.CreateAsync(nuGetProject, CancellationToken.None);
+                    var projectContextInfo = await ProjectContextInfo.CreateAsync(nuGetProject, CancellationToken.None);
                     context.Projects = new[] { projectContextInfo };
                     var nuGetUI = (NuGetUI)uiService;
                     nuGetUI.Projects = new[] { nuGetProject };
